Move utils exec/run handlers into new commands package
