+++ commands/run_test.go
@@ -1,4 +1,4 @@
-package utils
+package commands
 
 import (
 	"testing"
