+++ commands/testdata/test.sh
